html/semantics/embedded-content/media-elements/audio_controls_values.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Audio Test: audio_controls_values</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/2012/WD-html5-20121025/media-elements.html#htmlmediaelement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<audio id="m">
  Your browser does not support media elements.
</audio>
<div id="log"></div>
<script>
var media = document.getElementById("m");

test(function () {
  assert_false(media.controls, "media.controls is false");
}, "Check if the audio.controls returns false when it is absent");

test(function () {
  media.controls = null;
  assert_false(media.controls, "media.controls is false");
}, "Check if the audio.controls returns false when set it null");

test(function () {
  media.controls = "";
  assert_false(media.controls, "media.controls is false");
}, "Check if the audio.controls returns false when set it empty string");
</script>
